Add settings flyout
Settings flyout is added automatically to the workspace by `ffau.renderBlockly()`. An extra argument is added to `ffau.renderBlockly()` - the `settings` argument, which takes an array of objects containing info about what to include in the settings box. Currently, this is not yet parsed/inserted into the box, but it is accepted. All animation is also active with jQuery-UI.